Added support for DateTime
ResultSets still return the outdated `Date` or `Timestamp` types.  The only way
to get a modern timestamp is to use `getObject(column, OffsetDateTime.class)`
against a column of type `TIMESTAMP WITH TIME ZONE`.  We can't extend
ResultSet (in Java; for Kotlin we could provide an extension function).
However, we can add `SQLBuilder.getDateTime` which at least simplifies that for
the single-value shortcut.