Added support for enforcing tags matching method names
One of the initial ideas of tags in mock resultsets was to allow controlling if the mock
resultset is consumed by the correct caller. So far, this inspection could only be done
manually (e.g. using a breakpoint in MockSQLBuilderProvider::getRS).  This commit now
adds an option to enforce that these tags actually match the method name. If enabled,
MockSQLBuilderProvider will throw an IllegalArgumentException for mismatches.

Sometimes, methods have more than one SQLBuilder consumer.  To handle such cases, the
tag can be written as "<method name>:<tag>".  The enforcement will only look at the
part before the first ":".

This also fixes a test case issue where we disabled the "autogenerate mock resultset"
feature in a test case but forgot to restore it, causing subsequent tests to fail.